Introduce contacts and link to players.
Instead of putting contact details into the player record the model has now changed to have a contact table and a many to many link between players and contacts.

The relationship table has an additional field named is_primary to allow a contact to be marked as the primary contact.

The player list now shows the email and phone in the table associated with the primary contact only.